+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/skins/MarkerNodeSkin.kt
@@ -24,7 +24,7 @@ import javafx.scene.layout.VBox
 import org.kordamp.ikonli.javafx.FontIcon
 import org.kordamp.ikonli.material.Material
 import org.kordamp.ikonli.materialdesign.MaterialDesign
-import org.wycliffeassociates.otter.jvm.controls.MarkerNode
+import org.wycliffeassociates.otter.jvm.controls.marker.MarkerNode
 import tornadofx.*
 
 class MarkerNodeSkin(val control: MarkerNode) : SkinBase<MarkerNode>(control) {
